backend/requirements.txt
Flask>=3.0,<4.0
gunicorn>=21.2
peewee>=3.17
requests>=2.31
python-dotenv>=1.0
